chore: move codegen and device descriptors into @isomorphic
- Move server/codegen and the recorder actions.d.ts into
  packages/isomorphic/codegen
- Move deviceDescriptors.ts and deviceDescriptorsSource.json into
  packages/isomorphic
- Update references, DEPS.list files, and build/doc scripts
