+++ packages/injected/src/recorder/pollingRecorder.ts
@@ -18,7 +18,7 @@ import { Recorder } from './recorder';
 
 import type { InjectedScript } from '../injectedScript';
 import type { RecorderDelegate } from './recorder';
-import type * as actions from '@recorder/actions';
+import type * as actions from '@isomorphic/codegen/actions';
 import type { ElementInfo, Mode, OverlayState, UIState } from '@recorder/recorderTypes';
 
 interface Embedder {
