+++ packages/isomorphic/codegen/DEPS.list
@@ -1,0 +1,1 @@
+[*]
