+++ packages/isomorphic/codegen/language.ts
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-import type { BrowserContextOptions } from '../../..';
-import type * as types from '../types';
+import type { BrowserContextOptions } from 'playwright-core';
+import type * as types from './types';
 import type { LanguageGenerator, LanguageGeneratorOptions } from './types';
-import type * as actions from '@recorder/actions';
+import type * as actions from './actions';
 
 export function generateCode(actions: actions.ActionInContext[], languageGenerator: LanguageGenerator, options: LanguageGeneratorOptions) {
   const header = languageGenerator.generateHeader(options);
