+++ packages/isomorphic/codegen/python.ts
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-import { asLocator } from '@isomorphic/locatorGenerators';
-import { escapeWithQuotes, toSnakeCase } from '@isomorphic/stringUtils';
+import { asLocator } from '../locatorGenerators';
+import { escapeWithQuotes, toSnakeCase } from '../stringUtils';
 import { sanitizeDeviceOptions, toClickOptionsForSourceCode, toKeyboardModifiers, toSignalMap } from './language';
 import { deviceDescriptors } from '../deviceDescriptors';
 
 import type { Language, LanguageGenerator, LanguageGeneratorOptions } from './types';
-import type { BrowserContextOptions } from '../../..';
-import type * as actions from '@recorder/actions';
+import type { BrowserContextOptions } from 'playwright-core';
+import type * as actions from './actions';
 
 export class PythonLanguageGenerator implements LanguageGenerator {
   id: string;
