+++ packages/playwright-core/src/server/DEPS.list
@@ -5,7 +5,6 @@
 ../package.ts
 ../protocol/
 ./
-./codegen/
 ./har/
 ./recorder/
 ./registry/
