+++ packages/playwright-core/src/server/debugController.ts
@@ -19,11 +19,11 @@ import { parseAriaSnapshotUnsafe } from '@isomorphic/ariaSnapshot';
 import { unsafeLocatorOrSelectorAsSelector } from '@isomorphic/locatorParser';
 import { gracefullyProcessExitDoNotHang } from '@utils/processLauncher';
 import { asLocator } from '@isomorphic/locatorGenerators';
+import { generateCode } from '@isomorphic/codegen/language';
+import { JavaScriptLanguageGenerator } from '@isomorphic/codegen/javascript';
 import { SdkObject, createInstrumentation } from './instrumentation';
 import { Recorder, RecorderEvent } from './recorder';
-import { generateCode } from './codegen/language';
 import { collapseActions } from './recorder/recorderUtils';
-import { JavaScriptLanguageGenerator } from './codegen/javascript';
 
 import type { Language } from '@isomorphic/locatorGenerators';
 import type { BrowserContext } from './browserContext';
@@ -31,7 +31,7 @@ import type { InstrumentationListener } from './instrumentation';
 import type { Playwright } from './playwright';
 import type { ElementInfo, Mode } from '@recorder/recorderTypes';
 import type { Progress } from './progress';
-import type * as actions from '@recorder/actions';
+import type * as actions from '@isomorphic/codegen/actions';
 
 export class DebugController extends SdkObject {
   static Events = {
