+++ packages/playwright-core/src/server/index.ts
@@ -17,7 +17,7 @@
 export { Browser } from './browser';
 export { BrowserContext } from './browserContext';
 export { findRepeatedSubsequencesForTest } from './callLog';
-export { deviceDescriptors } from './deviceDescriptors';
+export { deviceDescriptors } from '@isomorphic/deviceDescriptors';
 export { DispatcherConnection, RootDispatcher, setMaxDispatchersForTest } from './dispatchers/dispatcher';
 export { RequestDispatcher, ResponseDispatcher } from './dispatchers/networkDispatchers';
 export { PlaywrightDispatcher } from './dispatchers/playwrightDispatcher';
