+++ packages/playwright-core/src/server/recorder/DEPS.list
@@ -2,7 +2,5 @@
 @isomorphic/**
 @utils/**
 ../
-../codegen/language.ts
-../codegen/languages.ts
 ../../package.ts
 node_modules/mime
