+++ packages/playwright-core/src/server/recorder/recorderRunner.ts
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-import { toKeyboardModifiers } from '../codegen/language';
+import { toKeyboardModifiers } from '@isomorphic/codegen/language';
 import { buildFullSelector, mainFrameForAction } from './recorderUtils';
 import { Progress } from '../progress';
 
 import type { Page } from '../page';
 import type * as types from '../types';
-import type * as actions from '@recorder/actions';
+import type * as actions from '@isomorphic/codegen/actions';
 import type { Frame } from '../frames';
 
 export async function performAction(progress: Progress, pageAliases: Map<Page, string>, actionInContext: actions.ActionInContext) {
