+++ packages/playwright-core/src/server/recorder/recorderSignalProcessor.ts
@@ -19,9 +19,9 @@ import { isUnderTest } from '@utils/debug';
 import { generateFrameSelector } from './recorderUtils';
 import { nullProgress } from '../progress';
 
-import type { Signal } from '../../../../recorder/src/actions';
+import type { Signal } from '@isomorphic/codegen/actions';
 import type { Frame } from '../frames';
-import type * as actions from '@recorder/actions';
+import type * as actions from '@isomorphic/codegen/actions';
 
 export interface ProcessorDelegate {
   addAction(actionInContext: actions.ActionInContext): void;
